requirements.txt
fastapi
uvicorn
google-generativeai
PyPDF2
python-docx
python-dotenv
python-multipart